test: consolidate Vera Custer ability
